Automatically minify .json files in built .jar
As Minecraft uses .json files for many purposes, mods tend to contain many .json files. This addition to the build script minifies .json files in the built mod .jar file. It does not affect the "source" .json files.

(I've also taken the liberty of simplifying some of the processResources logic, to make it compatible with Gradle 7.)